+++ sptrans/doc.go
@@ -13,14 +13,14 @@ Authentication:
 
 Endpoints:
 
-Routes
+Lines
 
-	// Search by route description or number
-	client.Route.Search("Lapa")
-	client.Route.Search("8000")
+	// Search by line description or number
+	client.Line.Search("Lapa")
+	client.Line.Search("8000")
 
-       // Search by route direction
-       client.Route.SearchByDirection("Lapa", 1)
+       // Search by line direction
+       client.Line.SearchByDirection("Lapa", 1)
 
 Stops
 
@@ -28,8 +28,8 @@ Stops
        client.Stop.Search("Afonso")
        client.Stop.Search("Rua Baltharzar da Veiga")
 
-       // Search by route code
-       client.Stop.SearchByRoute(1273)
+       // Search by line code
+       client.Stop.SearchByLine(1273)
 
        // Search by corridor code
        client.Stop.SearchByCorridor(8)