+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## 0.1.0 - 2017-09-17
 ### Added
 - Project structure
-- Routes endpoints
+- Lines endpoints
 - Bootstrap documentation
 - godoc
 - goreport